+++ dart_fusion/lib/src/d_service.dart
@@ -46,11 +46,12 @@ class DService {
           } else {
             // [5] Return failed response
             final body = await response.body();
-            String message;
-            try {
-              message = jsonDecode(body).toString();
-            } catch (e) {
-              message = body;
+            bool isHavingMessage() {
+              try {
+                return (jsonDecode(body) as JSON).isNotEmpty;
+              } catch (e) {
+                return body.isNotEmpty;
+              }
             }
 
             throw ResponseException(
@@ -58,7 +59,7 @@ class DService {
                 statusCode: response.statusCode,
                 body: ResponseModel(
                   message: DParse.httpStatusMessage(response.statusCode) +
-                      (message.isNotEmpty ? '\n\n --- \n\n$body' : body),
+                      (isHavingMessage() ? '\n\n --- \n\n$body' : body),
                 ).toJSON,
               ),
             );